backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Real-Time Microphone Transcription API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "faster-whisper>=1.1.0",
    "numpy>=1.26.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.9",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "pyjwt>=2.8.0",
    "argon2-cffi>=23.1.0",
    "itsdangerous>=2.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
]
